backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
google-generativeai==0.3.2
reportlab==4.0.4
ebooklib==0.18
markdown==3.5.1
python-dotenv==1.0.0
supabase>=2.8.0
gunicorn==21.2.0
PyJWT==2.8.0